Split out WiremockEmrServerlessTestHelper
